requirements.txt
langchain
sentence-transformers
transformers
torch
numpy
faiss-cpu
pypdf
tqdm
